Refactors fix for advanced search not clearing.
REF BL-141

Description
===
The current solution to BL-141 has some duplication and will need to be
modified if ever we add or remove fields in the advanced search section.
This commit makes the following changes:

* It refactors the code to make it more resistant to change and
also remove some of the duplication.

* It adds back labels for OR, and AND ops labels that were missing from
the buttons.

* It adds some new tests to this area of the code

QA Steps
==
- [ ] After running and advanced search, clearing specific sections of
the search should continue to have the expected behavior.