Fix error undefined [] for nil class.
REF BL-95

Description
===

The current implementation of the #guided_search method incorrectly
assumes that #search_field_def_for_key returns a hash even when a field
is not found for a specific key.  This change corrects that assumption
by coercing nil into a hash before attempting to use it, and also by
coercing nil into an empty string ("") in the case where :label is not
defined in the field definition.