Fixes range limit not plugin not working.
REF BL-205 BL-112 BL-27

When a user opens the "Date" facet, the range limit should work as
expected.  However, do to a controller mis-configuration an ajax call to
the catalog/range_limit route was throwing 500.

This commit updates the config/routes.rb removing duplications and
organizing it so that it's easier to read.

It also changes the solr range limit field to be pub_date_sort because
that field is indexed as a trint which is the desired format for the
range limit.

QA Steps
==
- [ ] Go to the root of the site.  Click on the "Date" facet, you should
see a graph with years under it. not a box with -1 and 1 on each side.
- [ ] Go to
http://localhost:3000/catalog/range_limit?range_end=2080&range_field=pub_date_sort&range_start=1100
you should see a list of dates and counts for each date.