Add a precision operator to advanced searches.
REF BL-133

  ## Description
Adds a precision operator to allow searches on :is, :begins_with and the
default :contains.

This change requires a Solr reindex to take effect.

Currently *_t fields are automatically flanked with `matchbeginswith`
and `matchendswith` on indexing.  On a :begins_with query the solr
parameters are updated to add "matchwithbegins foo bar" in the case of
a :begins_with `foo bar` search.

The :is (exact) phrase works in a similar way, by updating the
search value to be quoted instead of free.

 ## Additional changes
* I added a new container called "cli" to docker-compose, this adds an
editor which to containers in cases where we need to attach to the
container and edit something directly we don't have to always download
an editor first.

* I made it so that both cli and sh containers now use the base app
container in order to reuse that container image.