Use regular docker-compose up to set up app.
Rake docker:up is nice to have but all steps to set up the app should be
captured by docker-compose up command script.

This is so that we can assume a standard across projects.
If `docker-copose up` brings the site up with no additional steps
required (which is always possible) then that becomes quite a useful
interface.