Refactor advanced search preprocessors. (#264)
* Refactor advanced search preprocessors.

REF BL-253

The Blacklight API is very clear that the easiest way to customize
queries against the solr engine is by adding processors to the
default_processor_chain.  Unfortantely the blacklight_advanced_search
plugin uses this strategy by adding some processors that do not play
nice.

My origional strategy of adding a down stream processor was not working
because the blacklight_advanced_search plugin does not modify queries in
a consistant and predictive way (See comments on BL-253 for at least two
examples of this).

When I decided to refactor my procedures so that they came before
blacklight_advanced_search, I ran into yet another issue; the
blacklight_advanced_search processor chain procedures relie heavily on
blacklight_params.  And because I could not manipulate these parameters
from within an upstream chain procedure and hand the updated veresion to
the downstream advanced search procedure to use, I was left with two
obvious choices:

1. Override the blacklight_advanced_search procedure that was altering
the queries in order to pass in my custom changes to blacklight_params.

2. Override blacklight_params.

I created an issue on the blacklight_advanced_search site describing
this situation: projectblacklight/blacklight_advanced_search#82

I've gone with option 2 because it affords the least maintance on our
part.

This commit adds the following changes

(a) We removed all the code that was parsing the advanced_search query.
We'll just let that plugin do it's thing and manipulate it's behavior by
making updates to it's dependencies (specifically blacklight_params)

(b) We moved the changes that we require so that they happen before
blacklight_advanced_search does it's thing.

(c) We've added actual feature tests for the advanced_search behavior
because too many factors can affect the searching system.  By adding
feature tests at that level we minimize the possibility of quietly
breaking a functionality.

* Refactor .process_begins_with