Refactor harvest timing code.
REF BL-329

Once we broke up the harvest to happen for two different solr instances
we needed a way to run both jobs using the same parameters.  That means
we have to parameterize along the from and to options as well as the
solr URL.  Otherwise two consecutive calls to the job results in two
different sets of parameters.

This PR updates the task to allow us to parameterize the job as well as
adds a new task to allow us to capture the last build time in a way that
we can apply it to multiple Jenkins builds.