Refactor harvest timing code. (#282)
* Refactor harvest timing code.

REF BL-329

Once we broke up the harvest to happen for two different solr instances
we needed a way to run both jobs using the same parameters.  That means
we have to parameterize along the from and to options as well as the
solr URL.  Otherwise two consecutive calls to the job results in two
different sets of parameters.

This PR updates the task to allow us to parameterize the job as well as
adds a new task to allow us to capture the last build time in a way that
we can apply it to multiple Jenkins builds.

* Fix wrong oai provider URL.

* Tee logs to STDOUT for more verbosity to console.

REF BL-330

* Direct more logging to STDOUT.

REF BL-330

* Invoke the conform and ingest tasks.

REF BL-329

We need to invoke the task or it does not get run.

* Update error log file name.

* Clean up all files before new build.

Not all the process files where being cleaned up by the previous clean
up code.