Use default values in form (#286)
* Use default values in form.

REF BL-291

The advanced search form is not currently using the default values for
the operators which are available in the blacklight_params.

This commit fixes this issue, but does not fix an underlying issue which
I have discovered while working on this issue (BL-334).

* Only use latest values from op_row.

There is a bug in blacklight_advanced_search that appends new values to
previous values when the query parameter is of type array. Because
op_row is an array it grows for every search.  This commit guarantees
that we are using the latest values from the last commit.