Allow use of helper_method for index_fields.
REF BL-387

In order for helper methods to be automatically applied to a field value
must be rendered with the default <presenter>.field_value method.
However, field values rendered this way can be automatically escaped
which causes issues when the helper_method returns html.

This change updates the template so that it acts on a field[:raw]
setting that will allow for the helper_method to return raw html while
also rendering the field in the expected manner.