Fix next and previous document links for primo.
REF BL-386

When browsing a full view of a document, blacklight build a new query
that sets @rows and @stat attributes in the query builder instance.

These values need to be used in the query that is generated to get the
next and previous documents.  Prior to this fix the primo query builder
was defaulting to a limit of 10, and offset of 0.