Fix google_book image not showing up.
REF BL-470

I've studied the code for generating the google_book images and found
multiple  potential sources of issues that can cause images to not show
up when they otherwise might.

1. We are only selecting one id item from lccn or isbn number even when
multiple are present.  We are selecting the first item.  But this can
cause issues if the sort order is different in different system.

2. We are choosing either the lccn or isbn number but not both.

3. We are not properly handling cases where the id is returned as
a string and not a list of ids.

This change updates the book-image code to search for images on multiple
id and id types so that sorting is not an issue and so that if one id
fails we still try other ids that might return a book image.

This change also fixes handling the case where a string value is
returned to the function vs an array of strings.  In the case where
a string value was being returned we were selecting the first letter of
the id and using that first letter as an id for an image.