Fix not properly mapping fields in advanced search.
REF BL-398

Because we are not properly mapping fields in the advanced search
context we are seeing either failing searches or getting odd results.

This commit fixes that issue by making sure the filter is applied in the
advanced search context and adding a map from `description` to `desc`

TODO: Add add maps in Primo gem.