Do not use four_digit_year method unnecessarily. (#436)
When we were using 260 for date field there was some ambiguity about
where and what the four digit date was in the field.  Thus, a search of
the string was necessary.

However, now that we are using field 008, the four digit date is in
a very specific area per a standard.  Thus a search is not needed.

I'm not removing the method for now because it still may prove useful.
(But I'm also OK removing that method if that's a better choice).

@see https://www.loc.gov/marc/bibliographic/bd008.html