Add custom meta data to catalog email record.
REF BL-500

Unfortunately the way to add fields to emails and sms differs from the
standard way of adding fields in other contexts.

I've created a BL issue to address this. But because that work may
likely introduce some backwards incompatibility, I decided to implement
this change in the non standard way and I may in the future refactor
this work to add the fields via the catalog controller per change in BL.

see https://github.com/projectblacklight/blacklight/issues/1923