Bail on failed primo re-lookup. (#510)
* Bail on failed primo re-lookup.

REF BL-475

There is yet another bug with the Primo API that causes intermittent
failure of a search (i.e. the same search sometimes returns an item
sometimes returns no item.).  This commit updates the re-lookup code to
not use a nil doc in case of failure.

* Fetch item from individual record API.

REF BL-457

This seems to work better than finding a single item in the multi record
search.