Monitor articles not found.
REF BL-536

There have times where valid ids may not be returning articles from pnx
api.  We want to monitor these events to be able to report issue.

This commit updates the exception thrown when an article is not found so
that the controller does not gracefully handle the event and the
exception gets sent to honeybadger.