Do not cache the xhr format of sign_in form.
REF BL-565

The xhr format of the sign in form is the simple form we get in modal
context.  If the browser caches that before it caches the regular form,
then clicking the back button after logout from account page will bring
you back to a very simple page with a form in it (the modal context sign
in).

The expected behavior is that clicking the back button will bring you
the
regular sign in (non modal) page.  This is achieved by not allowing the
browser to cache the modal sign in page.