Only ingest once per build.
I'm tired of constantly waiting on my local environment to re-ingest all
the solr documents every time I run `docker-compose up app` even when
solr already has files in it.

This change updates the start-up script to check if docs are present and
only ingest if they are not.

A manual ingest step can be used to force an ingest if so desired.