Move field configuration back to controller.
Recently we moved out some field configurations to a module so that we could
share them  with the documents in the article context.

Although that works it is at the cost of adding  technical debt because
now there are two places to look for article field configurations instead of
one.

I noticed that when  a new response is created, BL injects the
blacklight_config object into the response.  So, I've decided to apply
the same technique to article documents.

Now when we create a new article document the configurations are injected into
the instance and thus available for use and no longer require the
document to be configurable or to load configurations from another file.