Force reload of referer if user gets auth error.
REF BL-576

There is a cross-site-request-forgery session token that gets added
to pages.  If a user signs in and then signs out and then tries to go
back (without reloading page) the same token appears and this causes
rails to raise an authentication exception if the user tries to follow
a link.

There are tricks to get around this involving updating the token using
javascript but this approach seems too risky since it's essentially
trying to circumvent a security measure by poking holes in it.

The best solution I could find that has no security issues is to simply
force a reload of the page the user is trying to work with.