Normalize the primo/pnxs id.
REF BL-571

Unfortunately we are currently using two separate APIs for the primo
search functionality and they are not entirely comparable.  Thus when an
article is loaded in the show view its id is different than when we see
it in the index view.

Also, when we bookmark an article we use its id to bookmark it.
Therefore the id saved when we bookmark an article from the index view
must be the same as when we bookmark the same article from the show
view. Currently these ids are not the same.

Fortunately the id for the primo/pnxs and primo/search APIs seem to only
differ in that the latter do not have TN_ in the beginning of the id.

This commit normalizes the id that is used for the record document so
that regardless of whether the item is bookmarked from the index view or
the show view it will be the same as for the primo/search id.

This change also adds a short circuit to the code that determines what
items need to be loaded via ajax.