Fix bookmark tools not rendering all actions.
Because we are passing @document instead of @document_list to
the `render_show_actions` method in the bookmarks context, not all
actions can be added via configuration.

This change updates the bookmarks action template to pass @document_list to
`render_show_actions` and removes adding RIS action via the template and
instead adds that via configuration as would be expected.