Adds a way for admins to experiment with query tweaks.
REF BL-660

This is an experimental feature that toggles on by setting the
SOLR_SEARCH_TWEAK_ENABLE environment variable to "on". When enabled this
adds a link to the solr search forms named "Tweak Solr Query Fields".

When clicked, the link expands a set of links that each can be clicked
to reveal a textarea with the value for the field that will be passed to
solr.

Updates to any field value will affect the current search and all other
searches until a completely new search is started (i.e. start over is
clicked)

This is a relatively bare bones feature.  Enhancements may be added in
the future if it proves useful.