Do not try pass unknown fields. (#732)
REF BL-698

blacklight_range_limit will usurp the search_field value under certain
circumstances.  When we pass this value to primo, then the primo server
throws an error because that field is unknown. This change vets the
fields that are submitted and only allows those that have been
configured via blacklight_config.