BL-667 pickup locations by item (#736)
Our former logic was only working at the bib level, so patrons could not 
request a specific volume that was not available at their library.  This 
refines the current valid pickup locations to get rid of a bug that was 
putting in locations that shouldn't be there.  Adds logic to filter 
pickup locations by description at the item level.