Have make_date helper respect timezones
Dates coming back from Alma's api are in UTC. We need to explicitly tell
rails to use a timezone or strftime will output UTC times.

This fixes human readnable dates on the account page for due dates and
Renewal due dates when a user renews an item.

Sets the Application timezone to "Eastern Time (US & Canada)".