[BL-685/BL-712] Do not share facets across searchs. (#773)
REF BL-712
REF projectblacklight/blacklight#1989

 ## Issue Description
After some QA testing, Jackie and I were talking further about the persistence of filters between Articles and the Solr menu options, and realized that there were some issues with this. For instance, even when there is a field match (author, topic, resource type), each index uses different vocabularies such that the values frequently don't match, leading to no results. We also felt that it might be confusing if some filters persisted (ex. date) while others didn't. So as a result, we'd like to scale this back and have just the basic query persist between Articles and the other options. Even in cases where the user has switched the basic query for "All Fields" to something else (ex. Title), it will revert back to All Fields. The relationship between Books, Journals, and More will remain as is, however.

 ## Implementation Description
This issue is a result of the upstream code being
promiscuous with respect to allowing any facet like user parameter to
be turned into a facet filter for the query.  Thus, it is not possible
to have multiple facet url parameters that should not be shared with
all searches.

By adding a check at the search build level that skips adding a facet
if the facet field in the parameter is not known (via configuration),
 then multiple searches can co-exist in the same URL and not clash.

An upstream PR has been submitted to fix this issue at the root,
hopefully it will be accepted.