[BL-723, BL-340] Fix purchase order email not working. (#792)
REF BL-723, BL-340

 ## Description:

The new solr doc types cannot be emailed or texted because they are missing an include.  This is reproducible by doing a search in books or journals and clicking on a found item and then trying to email or text it.

There is also an issue with the purchase_order email template that this
fixes.