Refactors search relevancy tests.
As a developer I want to quickly decipher the reason that a test fails
in order to get started on fixing the underlying issue.

This commit adds two new rspec matchers that encapsulate some analysis
utils for the search results.

This commit also updates the tests to use the new utils as well as
refactor the tests themselves to be more declarative and avoid the extra
step of looking through a fixture file for expected items.