[BL-359] Fix bound-with doc throws on fetch.
The mms_id and the doc_id are not interchangeable for bound-with items.
Thus, when fetching a bound-with item using mms_id, solr will return nil
causing havoc.  This commit fixes the issue by passing doc_id along to
the controller as well as rescuing from error with a dummy Solr document
if all else fails.