[DevOps] Default last build time to yesterday.
A recent update in Jenkins forced me to update the password I use to
interact with Jenkins via its API.

I then forgot to update the file in lurch that keeps the shell
secrets for authentication. In turn, the API call to get the last
successful build time failed and we ended up using the default time
which was set to be 0.

This commit updates the code so that it fails outright instead of trying
to rescue with default assumptions.