[BL-779] Log API requests even on failure.
REF BL-779

When we make API calls we log the response process in order to obtain
metrics on these API requests.  However when the request raises an
exception we fail to log the event.  We need to make sure that logging
happens even in the event of a raised exception.

This change updates the `json_request_logger` method to accept a block.
When a block is passed it gets yielded before logging happens and if an
exception is raised it is caught, logged and then released so that we do
not interrupt the proper behavior of the app just for logging.