[DevOps] Do not crash container on ingest.
When the container is first tarting we have a script to populate solr
with data if no data is found. When this ingest step fails at start it
causes the container to stop. This commit wraps the ingest step so that
failure doesn't kill the app and allows us to use the app to investigate
the issue.