[BL-780] Redirect back to specific document on log in.
REF BL-780

When forced to log in in order to make a request the redirect link is not
document specific.  Therefore, when you are logged back in you get
redirected with focus on top of page.  This forces user to scroll back
down if they redirected from a lower section of the window.

This update adds target specific redirect links so that the browser scrolls
to the correct part of the window on log in.