[BL-666] Do not override standard advanced link behavior.
* Removes overriding template and helper method for articles advanced
link

We have started to use the `render_advanced_search_link` to handle all
the advanced_search_link cases.

However, the advanced articles link was being overridden via the template with
it's own specific helper.  Although that is arguably a better approach
(from a good OOP technique view) it's much better to just apply the same
technique to all our advanced links so that we do not inadvertantly
forget that some are implemented differently than others.