[DevOps] Add default solr timezone when not present.
At one point the Marc XML exports from Alma did not have a timezone set in the update times etc.  But, this changed and caused some failures do to time differences.

This commit adds a default timezone now that Solr is configured to expect it when parsing some time fields to guard against the possibility of Alma reverting to not printing time zones in it's time fields.

The default timezone is chosen to be Etc/UTC because that was the default from when Alma was not printing Time Zones.  The assumption being that if a switch thus happen that the switch is to that original state.