[BL-814] Fix issue with book bookmark items.
REF BL-814

 ## Description:

When items are bookmarked from the book or journal index pages they
do not appear as being bookmarked when viewing the bookmarked items.

This happens because of a conflict between the document.class and
bookmark.class.  For now we can fix this by just not checking against
that field as all our Solr document ids are uniq by default.