[BL-814] Fix bookmark not working from book or journal.
REF BL-814
REF BL-822

Having books and journal documents be represented by different models
SolrBookDocument and SolrJournalDocument that derive from the
SolrDocument is causing both the bookmark and email functionality for
these documents to fail when actions occur from the bookmark or journal
index pages.

This is partly do to a misconfiguration.  Although we need these models
in order to have correct paths .. they should be equal to SolrDocument
not derive from.

Updating them to equal SolrDocument fixes both the issues with
bookmarking and the issue with sending emails.

 ## Depends on

* Also adds temporary rake task to fix old data in the bookmarks table.
- [ ] This rake task needs to be run in order for fix to take full effect