[BL-833] Fix everything search fails for some queries.
When searching for "Development and validation of the children’s
constructive conflict resolution scale" on libqa or production an error
is thrown due to no method found.

We are fixing this by removing the call to this method from a shared
template.  The call is not required since nothing happens even if it is
evaluated.