[BL-847] Add config entry for DSC. (#1078)
There is an error on production that happens when availability item is on DSC:

https://librarysearch.temple.edu/catalog/991036860491803811

There are a couple of tasks that need to be done.
* Set a configuration for DSC
* Set a default if config is not available for DSC
* Notify Honeybadger when config not set (so that we can update configs)

Update for what's on master.

Orig commit was for v0.6

Don't bother with location_name_from_short_code method.

This method is different enough from v0.6.6 that there might not be any
issues.