[BL-809] Fix database title missing from results.
We recently removed automatically adding title_truncated_display to the
solrdocument instance.  Because databases inherits from the catalog
controller and also uses the solr document model, it was automatically
configured to use title_statement_display as the title field for
results.

There are two ways to fix this issue.
* Add a databases specific title field (like title_display)
* Or, make sure we add "title_truncated_display" to databases.

I when with the latter option because there is another related issue
that requires we truncate the database titles in the results page.