[BL-917] Rename az facet fields. (#1259)
Rename the az facet fields to avoid collision with the catalog faceting.

The catalog is aware of the facet fields that it has configured but in
order to work it cannot share the same name as facets for other
searches.

This commit updates the az facet fields and takes advantage of a BL
feature that allows us to change the name that BL uses for this field
without needing to change what the field is named in the backend solr
core (i.e. solrcores can duplicate field names).