[BL-886] Do not add more links for electronic resources. (#1268)
REF BL-886

When a Marc record has PRT data fields then it is an electronic
resource.  For these documents the resource link would be the same as the
more links because the resource link would always be generated from the
contents of the PRT data field.

This commit will remove the url_more_links for the record is an
electronic resource.