Add feature flag for campus closure (#1984)
Adds a feature flag that can toggle functionality that depends if our campus as open as ususal,  or closed for long periods of time because of social distancing requirements.

The feature flag uses the env var $CAMPUS_CLOSED, defaulting to false if the env var does not exist. If env var is set, will only evaluate to false if the value is "false". All others values evaluate to true.

The env var flag can be overriden by a URL paramter "campus_closed". If the URL parameter is present passed and the value is "false", the flag will evaluate as false. Any other value passed to the URL paramter will icause the flag to be evaluated as true.

Adds the class `FeatureFlags` with the method `campus_closed?` so that it can be called from anywhere in the code base.

Adds a convenience helper_method to the application_controller so that it can be called from controller or views with the the controller `params` hash passed as a paramater.