Feature flag some functionality
Adds inline js to the main layout to set campus_closed var

This adds a single line of javascript in the head that sets the js var `campus_closed` so that other js functions in the browser may access it. Needs to be rendered with each request so that flag maybe be overriden by url param.

Updates severl pieces of functionality to be feature flaggable.

UPdates some text.