BL-1310: Reenable article bookmark deletes.
This adds back bookmark options for articles in the bookmarks context.

This also fixes a bug with bookmark deletes do to pnx to cdi id
conflicts.

Finally, and unfortunately,  I had to also revert some changes we made
earlier that removed deprecation warnings.  However, without the revert
there is a bug that we did not notice because we had disabled article
bookmarks prior to merging that change.

Since we will be removing the custom code that was required to enable
article bookmarks, I think that it's OK to revert this one change as we
will fix the issue soon simply by removing the deprecated code.