BL-1540: Fix lc_classification constraint broken
A bug in the blacklight_range_limit gem was allowing the
lc_classification facet constraint to show up even though it wasn't been
configured as a facet field in the controller.

When this bug was fixed it caused the lc_classification constraint to
disappear.

This commit configures the lc_classification constraint but then removes
it from the search so that Solr does not throw an error since
"lc_classification" is not a field that our Solr config knows about.

This commit also adds a feature test for these constraint to avoid
a regression bug.