Allow adding javascript enabled feature specs.
We have some features that require JavaScript to test them properly.

However, JavaScript testing frameworks are not easy to manage and
understand when you are not constantly writing JavaScript.

A different approach we can take is to test these features via a web
browser interaction using a feature test that interacts with the codes
behavior.

This should suffice to test features in our site that require
JavaScript.

This change updates our code to:
* Add dependencies to the test and developer environment to enable
  JavaScript enabled feature specs.
* Adds an example test that test that clicking on facet headers opens
  the closed facets.